refactor: separate frontend directories (#22)
