+++ frontend/src/features/skillChaining/categories/filters/SelectedCategoryTags.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { getCategoryColor } from '../../utils/categoryColors';
+import { getCategoryColor } from '../hooks/categoryColors';
 
 interface SelectedCategoryTagsProps {
   selectedCategories: string[];
