+++ frontend/src/features/skillChaining/categories/hooks/categoryColors.ts
