+++ frontend/src/features/skillChaining/skills/components/SmartSearch.tsx
